==> 00_project-doc.Rmd <==
---
title: "Project Documentation"
output: html_document
# output: pdf_document
---

# Summary

The main contributions are based on sets of experiments structured in 3 sections. The synthetic data (section 3 and 4)
experiments are intended to (1) highlight specific problems often reported to occur in cluster analysis and how
topological manifold learning can solve these and (2) to show the subtleties of applying manifold learning as a tool to infer topological data structures. The real data experiments (section 4) are intend to substantiate the findings based on synthetic data on real world problems. Additional results are reported in the appendices 1 - 3. 

***Important Note***  
Some data files are too large to be uploaded to GitHub and need to be downloaded from https://syncandshare.lrz.de/getlink/fiSaZPAV3jzvGjAikmDQM3JS/ before `00_project-doc.Rmd` can be 
run!

Files to be placed in folder `data`:

  - `embs_2D_layout_real_dat.RData`
  - `embs_real_dat.RData`
  - `sec4_real_dats.RData` 
  
Files to be placed in folder `data/datasets`

  - `mnist_train.csv`


# Technical aspects
This document gathers all necessary information for reproducibility. In the default setup, knitting the document takes 
a couple (~ 10) minutes. All relevant code will be part of the output, computation heavy experiments are -- however --
not executed. The corresponding code chunks are set to `eval=FALSE` and provide rough estimates for computation times 
for the individual experiments (obtained on a standard notebook with Linux Mint 19.2 Cinnamon, 1.90GHz × 4, 16 GB RAM, 
8 cores). To run individual experiments the setup chunk needs to be run!

```{r setup, message=FALSE}  
knitr::opts_chunk$set(
  message = FALSE,
  warning = FALSE
)

source("R/setup.R")
source("R/help_funs.R")
```

# Section 3

```{r child = '01_section_3.Rmd'}
```

# Section 4

```{r child = '02_section_4.Rmd'}
```

# Section 5

```{r child = '03_section_5.Rmd'}
```

# Appendix A

```{r child = '04_appendix_samplvar.Rmd'}
```

# Appendix B

```{r child = '05_appendix_embvar.Rmd'}
```

# Appendix C

```{r child = '06_appendix_fgraph.Rmd'}
```

# Appendix D

```{r child = '07_appendix_embplots.Rmd'}
```
