==> 05_appendix_embvar.Rmd <==
## B1 Synthetic datasets $E_{100}$, $E_{1000}$, $U_{3}$ & $U_{1003}$

```{r sec2-embedding-variability, eval = FALSE}
# embed each of the 4 settings 25 times to assess effect of SGD in constructing embedding
# layout on cluster results
reps <- 25
set.seed(110)
rand_states_rep <- sample(1:10000000, length(exp_settings) * reps) 

rep_combs <- cbind.data.frame(
  rand_state = rand_states_rep, 
  setting = rep(names(exp_sec2_dists), each = reps)
)
  
# ~ 1 min
if (Sys.info()[["sysname"]] == "Windows") {
  cores_to_use <- 1
} else {
  cores_to_use <- 4
}

exp_sec2_emb_rep_dists <- 
  parallel::mcMap(
    function(set, rand) {
      emb <- umap::umap(
        as.matrix(exp_sec2_dists[[set]]), 
        random_state = rand, 
        n_neighbors = k, 
        n_components = d, 
        input = "dist")
      emb_dists <- dist(emb$layout)
      emb_dists
    },
    set = rep_combs$setting,
    rand = rep_combs$rand_state,
    mc.cores = cores_to_use
)

eps_range_rep <- seq(0.01, 15, by = 0.01)

# ~ 8 h with eps_range_rep = seq(0.01, 15, by = 0.01)
if (Sys.info()[["sysname"]] == "Windows") {
  cores_to_use <- 1
} else {
  cores_to_use <- 5
}

exp_sec2_rep_res <- parallel::mclapply(
  exp_sec2_emb_rep_dists,
  function(dat) cluster_res(dat, eps_range = eps_range_rep, lbls = lbls),
  mc.cores = cores_to_use
)

dt_sec2_rep_res <- cbind(
  rep = rep(seq_len(reps), each = length(eps_range_rep) *  length(exp_settings)),
  setting = rep(rep_combs$setting, each = length(eps_range_rep)),
  eps = rep(eps_range_rep, length(exp_sec2_rep_res)),
  as.data.table(do.call(rbind, exp_sec2_rep_res))
)

# save(dt_sec2_rep_res, file = "vignettes/data/res_sec2_rep.RData")
```

```{r embedding-variability-plots, echo=FALSE, fig.width = 7.5, fig.height = 10}

# DBSCAN applied directly to the data: optimal ARI/NMI and corresponding eps values
load("data/results/res_sec2_long-range.RData")

# dt_sec2_res_long[method == "DBSCAN", .("maxP" = max(Performance)), by = c("setting", "Measure")]

dt_temp <- dt_sec2_res_long[method == "DBSCAN"]

dbs_opt <- copy(
  dt_temp[, 
         .(maxPerformance = round(max(Performance), 2)),
         by = c("setting", "Measure")]
)

dbs_opt_eps <- copy(
  dt_temp[
    dt_temp[, 
             .I[which(Performance == max(Performance))], 
             by = c("setting", "Measure")]$V1
  ]
)



# Embedding replication results
load("data/results/res_sec2_rep.RData")

dt_sec2_rep_res[, c("rARI", "rNMI") := list(round(ARI, 3), round(NMI, 3))] 

dt_ARI <- dt_sec2_rep_res[, 
                          .(maxARI = max(rARI), meanARI = mean(rARI), minARI = min(rARI)), 
                          by = c("setting", "eps")]
dt_NMI <- dt_sec2_rep_res[, 
                          .(maxNMI = max(rNMI), meanNMI = mean(rNMI), minNMI = min(rNMI)), 
                          by = c("setting", "eps")]

dt_sec2_rep_ARI_long <-
  melt(dt_ARI, 
       id.vars = c("setting", "eps"), 
       measure.vars = c("maxARI", "meanARI", "minARI"), 
       variable.name = "Measure", 
       value.name = "Performance")

dt_sec2_rep_ARI_long[, 
                      setting := ordered(setting, 
                                         levels = c("e100", "e1000", "u3", "u1003"),
                                         labels = c("E100", "E1000", "U3", "U1003"))]

dt_sec2_rep_NMI_long <-
  melt(dt_NMI, 
       id.vars = c("setting", "eps"), 
       measure.vars = c("maxNMI", "meanNMI", "minNMI"), 
       variable.name = "Measure", 
       value.name = "Performance")

dt_sec2_rep_NMI_long[, 
                     setting := ordered(setting, 
                                        levels = c("e100", "e1000", "u3", "u1003"),
                                        labels = c("E100", "E1000", "U3", "U1003"))]

dbs_opt[, 
        setting := ordered(setting, 
                           levels = c("e100", "e1000", "u3", "u1003"),
                           labels = c("E100", "E1000", "U3", "U1003"))]

dbs_opt_eps[, 
            setting := ordered(setting, 
                               levels = c("e100", "e1000", "u3", "u1003"),
                               labels = c("E100", "E1000", "U3", "U1003"))]

plt_rep_ARI <- 
  ggplot(dt_sec2_rep_ARI_long) + 
    geom_line(aes(x = eps, y = Performance, color = Measure, lty = Measure)) +
    facet_wrap(~ setting, scales = "free_x", ncol = 1) +
    xlab(eps_tex) +
    ylim(0, 1) +
    # adding optimal ARI value of DBSCAN directly applied to the data has horizontal lines
    geom_hline(
      data = dbs_opt[Measure == "ARI", c("setting", "maxPerformance")], 
      aes(yintercept = maxPerformance), size = 0.25, lty = 2) +
    # adding corresponding eps-value of DBSCAN directly applied to the data has  vertical lines
    geom_vline(
      data = dbs_opt_eps[Measure == "ARI", c("setting", "eps")],
      aes(xintercept = eps), size = 0.25, lty = 2)


plt_rep_NMI <- 
  ggplot(dt_sec2_rep_NMI_long) +
    geom_line(aes(x = eps, y = Performance, color = Measure, lty = Measure)) +
    facet_wrap(~ setting, scales = "free_x", ncol = 1) +
    xlab(eps_tex) +
    ylim(0, 1) +
    # adding optimal NMI value of DBSCAN directly applied to the data has horizontal lines
    geom_hline(
      data = dbs_opt[Measure == "ARI", c("setting", "maxPerformance")], 
      aes(yintercept = maxPerformance), size = 0.25, lty = 2) +
    # adding corresponding eps-value of DBSCAN directly applied to the data has  vertical lines
    geom_vline(
      data = dbs_opt_eps[Measure == "ARI", c("setting", "eps")],
      aes(xintercept = eps), size = 0.25, lty = 2)

plt_rep_ARI + plt_rep_NMI + plot_layout(guides = "collect") & theme(legend.position = "bottom")
# ggsave("vignettes/figures/app_syn_reps.png", device = "png", width = 7.5, height = 10)
```

## B1 Real data

```{r real-rep, eval=FALSE}
load("data/sec4_real_dats.RData")

## Setup
reps <- 25
dat <- c("iris", "wine", "coil")
l_dats <- l_dats[dat]

# computation time ~ 5 min
for (i in seq_along(l_dats)) {
  l_dats[[i]]$dat_X <- as.matrix(dist(l_dats[[i]]$dat_X))
}

## UMAP
k <- c(5, 10, 15)
rand_state <- 412745030 + 100000 * (1:(reps * length(dat) * length(k)))
d <- 3

## DBSCAN
eps_range <- seq(0.01, 25, by = 0.01)


# Embeddings
emb_combs <- expand.grid(seq_len(reps), k, dat) # rep(k, each = reps)
colnames(emb_combs) <- c("reps", "k", "dat")
emb_combs <- cbind(emb_combs,
                   rand_state = rand_state)

if (Sys.info()[["sysname"]] == "Windows") {
  cores_to_use <- 1
} else {
  cores_to_use <- 5
}

# computation time ~ 5 min
l_embs <- parallel::mcmapply(
  function(dat, k, rs) umap::umap(
    l_dats[[dat]]$dat_X, 
    n_neighbors = k, 
    n_components = d, 
    random_state = rs,
    input = "dist"
  )$layout,
  dat = emb_combs$dat,
  k = emb_combs$k,
  rs = rand_state,
  SIMPLIFY = FALSE,
  mc.cores = cores_to_use
)

# Clusterings
clust_combs <- expand.grid(eps_range, seq_along(l_embs))
colnames(clust_combs) <- c("eps_val", "embedding_iter")

# computation time ~ 1h
m_res <- parallel::mclapply(
  seq_len(nrow(clust_combs)),
  function(iter, ...) {
    eps_val <- clust_combs$eps_val[iter]

    emb_iter <- clust_combs$embedding_iter[iter]
    emb_layout <- l_embs[[emb_iter]]

    dat <- emb_combs$dat[[emb_iter]]
    dat_lbls <- l_dats[[dat]]$dat_lbls

    clustering <- dbscan(emb_layout, eps = eps_val)
    perfm <- performance(clustering$cluster, dat_lbls)
  },
  mc.cores = cores_to_use
)

dt_res_real_rep <- cbind(
  data = rep(emb_combs$dat, each = length(eps_range)),
  k = rep(emb_combs$k, each = length(eps_range)),
  rep = rep(emb_combs$reps, each = length(eps_range)),
  eps = clust_combs$eps_val,
  as.data.table(do.call(rbind, m_res))
)
# save(dt_res_real_rep, file = "vignettes/data/res_sec4_rep.RData")
```

```{r real-rep-plts, echo=FALSE, fig.width = 9, fig.height = 14}
# DBSCAN applied directly to the data: optimal ARI/NMI and corresponding eps values
load("data/results/res_real_full.RData")

dt_res_real_long <- melt(
  dt_res_real_full, 
  id.vars = c("Data", "Method", "eps", "k"), 
  measure.vars = c("ARI", "NMI"), 
  variable.name = "Measure", 
  value.name = "Performance"
)

dbs_temp <- copy(
  dt_res_real_long[Data %in% c("Iris", "Wine", "COIL") & Method == "DBSCAN", ]
)
setnames(dbs_temp, "Data", "data")
dbs_temp$data <- ordered(dbs_temp$data)

dbs_opt <- copy(
  dbs_temp[, 
           .(maxPerformance = round(max(Performance), 2)),
           by = c("data", "Method", "k", "Measure")]
)

dbs_opt_eps <- copy(
  dbs_temp[
    dbs_temp[, 
             .I[which(Performance == max(Performance))], 
             by = c("data", "Method", "k", "Measure")]$V1
  ]
)

# Embedding replication results
load("data/results/res_sec4_rep.RData")
dt_res_real_rep[, c("rARI", "rNMI") := list(round(ARI, 3), round(NMI, 3))] 
dt_res_real_rep[, k := factor(k, labels = c("UMAP-5", "UMAP-10", "UMAP-15"))]
dt_res_real_rep[, data := factor(data, labels = c("Iris", "Wine", "COIL"))]

dt_real_ARI <- dt_res_real_rep[, 
                               .(maxARI = max(rARI), meanARI = mean(rARI), minARI = min(rARI)), 
                               by = c("data", "k", "eps")]
dt_real_NMI <- dt_res_real_rep[, 
                               .(maxNMI = max(rNMI), meanNMI = mean(rNMI), minNMI = min(rNMI)), 
                               by = c("data", "k", "eps")]

dt_real_rep_ARI_long <-
  melt(dt_real_ARI, 
       id.vars = c("data", "k", "eps"), 
       measure.vars = c("maxARI", "meanARI", "minARI"), 
       variable.name = "Measure", 
       value.name = "Performance")

dt_real_rep_NMI_long <-
  melt(dt_real_NMI, 
       id.vars = c("data", "k", "eps"), 
       measure.vars = c("maxNMI", "meanNMI", "minNMI"), 
       variable.name = "Measure", 
       value.name = "Performance")


plt_real_rep_ARI <- 
  ggplot(dt_real_rep_ARI_long) + 
    geom_line(aes(x = eps, y = Performance, color = Measure, lty = Measure)) +
    facet_wrap(~ data + k, scales = "free_x", ncol = 3) +
    xlab(eps_tex) +
    ylim(0, 1) +
    # adding optimal ARI value of DBSCAN directly applied to the data has horizontal lines
    geom_hline(
      data = dbs_opt[Measure == "ARI", c("data", "maxPerformance")], 
      aes(yintercept = maxPerformance), size = 0.25, lty = 2) +
    geom_vline(
      data = dbs_opt_eps[Measure == "ARI", c("data", "eps")],
      aes(xintercept = eps), size = 0.25, lty = 2) +
  theme(legend.position = "bottom")

plt_real_rep_NMI <- 
  ggplot(dt_real_rep_NMI_long) +
    geom_line(aes(x = eps, y = Performance, color = Measure, lty = Measure)) +
    facet_wrap(~ data + k, scales = "free_x", ncol = 3) +
    xlab(eps_tex) +
    ylim(0, 1) +
    # adding optimal NMI value of DBSCAN directly applied to the data has horizontal lines
    geom_hline(
      data = dbs_opt[Measure == "NMI", c("data", "maxPerformance")], 
      aes(yintercept = maxPerformance), size = 0.25, lty = 2) +
    geom_vline(
      data = dbs_opt_eps[Measure == "NMI", c("data", "eps")],
      aes(xintercept = eps), size = 0.25, lty = 2) + 
  theme(legend.position = "bottom")

plt_real_rep_ARI / plt_real_rep_NMI + plot_annotation(tag_levels = "A")
# ggsave("vignettes/figures/app_real_reps.png", device = "png", width = 9, height = 14)
```
