==> 06_appendix_fgraph.Rmd <==

```{r fgraph-extention, eval=FALSE}
load("data/sec4_real_dats.RData")

## Setup
dat_selection <- c("iris", "wine", "coil", "pendigits")
l_dats <- l_dats[dat_selection]

# due the transformation d_ij = 1 - v_ij \in [0, 1], no eps values larger than 1 make sense.
# Since pendigits has many ob.s, reducing the eps range considerably reduces computation time.
# As the other datasets have lower number of obs, this is not as important.
l_dats$pendigits$eps_range <- seq(0.01, 1, by = 0.01)

min_dist <- 0.1 # note: not used in graph constr.
k <- c(5, 10, 15)

emb_combs <- expand.grid(k, dat_selection)
names(emb_combs) <- c("k", "data")

if (Sys.info()[["sysname"]] == "Windows") {
  cores_to_use <- 1
} else {
  cores_to_use <- 3
}

# fuzzy graphs for different values of k (only fgraph --> n_epochs = 0)
l_embs_fgr <- parallel::mcmapply(
  function(dat, k) uwot::umap(
    l_dats[[dat]]$dat_X,
    n_neighbors = k,
    ret_extra = "fgraph", 
    n_epochs = 0,
    min_dist = min_dist),
  dat = emb_combs$data,
  k = emb_combs$k,
  SIMPLIFY = FALSE,
  mc.cores = cores_to_use
)

# turn similarities into dissimilarities via 1 - v_ij
l_emb_fgr_dis <- lapply(
  l_embs_fgr,
  function(emb) {
    temp <- as.matrix(1 - round(emb$fgraph, 5))
    diag(temp) <- 0
    as.dist(temp)
  }
)

exp_combs <- 
  cbind.data.frame(
    setting = paste(emb_combs$dat, emb_combs$k, sep = "_"),
    dat = rep(names(l_dats), each = length(k))
  )
names(l_emb_fgr_dis) <- exp_combs$setting


if (Sys.info()[["sysname"]] == "Windows") {
  cores_to_use <- 1
} else {
  cores_to_use <- 3
}

# computation time ~ 45 min
exp_real_fgr_res <- parallel::mcmapply(
  function(set, dat) cluster_res(
    dat = l_emb_fgr_dis[[set]],
    eps_range = l_dats[[dat]]$eps_range,
    lbls = l_dats[[dat]]$dat_lbls
  ),
  set = exp_combs$setting,
  dat = exp_combs$dat,
  SIMPLIFY = FALSE,
  mc.cores = cores_to_use
)

eps_range_length <- sapply(l_dats[dat_selection], function(dat) length(dat$eps_range))

dt_res_real_fgraph <- cbind(
  data = unlist(lapply(
    dat_selection, 
    function(dat) rep(dat, length(k) * length(l_dats[[dat]]$eps_range)))
  ),
  k = rep(rep(k, length(dat_selection)), rep(eps_range_length, length(k))),
  eps = unname(unlist(
    lapply(l_dats[dat_selection],
           function(dat) rep(dat$eps_range, length(k))))),
  as.data.table(do.call(rbind, exp_real_fgr_res))
)

# save(dt_res_real_fgraph, file = "vignettes/data/res_app_fgraph.RData")
```

```{r fgraph-effect, echo =FALSE}

# real data with UMAP-fgraph only
load("data/results/res_app_fgraph.RData")

# real data with UMAP fgraph and embedding
load("data/results/res_real_full.RData")

# UMAP-fgr
dt_res_real_fgraph_long <- melt(
  dt_res_real_fgraph, 
  id.vars = c("data", "eps", "k"), 
  measure.vars = c("ARI", "NMI"), 
  variable.name = "Measure", 
  value.name = "Performance"
)

# set slightly negative ARI values to zero
dt_res_real_fgraph_long[Performance < 0, Performance := 0]

dt_res_real_fgraph_long[, c("data", "k", "Performance") := list(
  ordered(factor(data), 
          levels = c("iris", "wine", "coil", "pendigits"), 
          labels = c("Iris", "Wine", "COIL", "Pendigits")),
  ordered(factor(k), labels = c("UMAP-5", "UMAP-10", "UMAP-15")),
  round(Performance, 3))
]


# UMAP-fgr+emb
dt_res_real_long <- melt(
  dt_res_real_full, 
  id.vars = c("Data", "Method", "eps", "k"), 
  measure.vars = c("ARI", "NMI"), 
  variable.name = "Measure", 
  value.name = "Performance"
)

dt_res_real_long$Data <- ordered(
  factor(dt_res_real_long$Data), 
  levels = c("Iris", "Wine", "Pendigits", "COIL", "MNIST", "FMNIST-10", "FMNIST-5")
)

# compare eps ranges

## Iris & Wine
for (i in c("Iris", "Wine")) {
  opt_vals <- print(dt_res_real_long[Data == i & k == "10", max(Performance), by = "Measure"])$V1
  opt_vals_fgr <- print(dt_res_real_fgraph_long[data == i & k == "UMAP-10", max(Performance), by = "Measure"])$V1
  names(opt_vals) <- names(opt_vals_fgr) <- c("ARI", "NMI")
  
  cat("\n", "\n",
    i, "\n", "\n", 
    "fgraph+embedding\n\n",
    "Optimal values:\n",
    round(opt_vals, 2),
    "\n",
    "Optimal eps-range:\n",
    dt_res_real_long[Data == i & k == 10 & Measure == "ARI"][Performance == max(Performance), range(eps)],
    "\n",
    dt_res_real_long[Data == i & k == 10 & Measure == "NMI"][Performance == max(Performance), range(eps)],
    "\n\n",
    "fgraph only\n\n",
    "Optimal values:\n",
    round(opt_vals_fgr, 2),
    "\n",
    "Optimal eps-range:\n",
    dt_res_real_fgraph_long[data == i & k == "UMAP-10" & Measure == "ARI"][Performance == max(Performance), range(eps)],
    "\n",
    dt_res_real_fgraph_long[data == i & k == "UMAP-10" & Measure == "NMI"][Performance == max(Performance), range(eps)]
  )
}

## COIL & Pendigits
for (i in c("COIL", "Pendigits")) {
   cat("\n", "\n",
    i, "\n", "\n", 
    "fgraph+embedding\n\n",
    "Optimal eps-range (ARI/NMI > 0.6):\n",
    dt_res_real_long[Data == i & k == 10 & Measure == "ARI"][Performance > 0.6, range(eps)],
    "\n",
    dt_res_real_long[Data == i & k == 10 & Measure == "NMI"][Performance > 0.6, range(eps)],
    "\n\n",
    "fgraph only\n\n",
    "Optimal eps-range (ARI/NMI > 0.6):\n",
    dt_res_real_fgraph_long[data == i & k == "UMAP-10" & Measure == "ARI"][Performance > 0.6, range(eps)],
    "\n",
    dt_res_real_fgraph_long[data == i & k == "UMAP-10" & Measure == "NMI"][Performance > 0.6, range(eps)]
  )
}
```


```{r fgrph-plots, echo=FALSE, fig.height = 9, fig.width = 9}
load("data/results/res_app_fgraph.RData")

dt_res_real_fgraph_long <- melt(
  dt_res_real_fgraph, 
  id.vars = c("data", "eps", "k"), 
  measure.vars = c("ARI", "NMI"), 
  variable.name = "Measure", 
  value.name = "Performance"
)

# set slightly negative ARI values to zero
dt_res_real_fgraph_long[Performance < 0, Performance := 0]

dt_res_real_fgraph_long[, c("data", "k", "Performance") := list(
  ordered(factor(data), 
          levels = c("iris", "wine", "coil", "pendigits"), 
          labels = c("Iris", "Wine", "COIL", "Pendigits")),
  ordered(factor(k), labels = c("UMAP-5", "UMAP-10", "UMAP-15")),
  round(Performance, 3))
]

dt_temp <- copy(
  dt_res_real_fgraph_long[, 
                          .(maxPerformance = round(max(Performance), 2)), 
                          by = c("data", "k", "Measure")]
)

dcast(dt_temp, data ~ k + Measure, value.var = "maxPerformance") 

ggplot(dt_res_real_fgraph_long[eps <= 1.5]) +
    geom_line(aes(x = eps, y = Performance, color = Measure, lty = Measure)) +
    facet_wrap(~ data + k, scales = "free_x", ncol = 3) +
    xlab(eps_tex) +
    ylim(0, 1) +
    theme(legend.position = "bottom")

# ggsave("vignettes/figures/app1_res_fgraph.png", device = "png", height = 9, width = 9)
```
